--- components/common/CharmEditor/components/@bangle.dev/tooltip/index.ts
@@ -1,13 +1,0 @@
-import * as selectionTooltip from './selectionTooltip';
-import type { SuggestTooltipRenderOpts } from './suggestTooltip';
-import * as suggestTooltip from './suggestTooltip';
-
-export type { SelectionTooltipProps } from './selectionTooltip';
-
-export * as tooltipPlacement from './tooltipPlacement';
-export type { TooltipRenderOpts } from './tooltipPlacement';
-export { selectionTooltip };
-export { createTooltipDOM } from './createTooltipDOM';
-
-export { suggestTooltip };
-export type { SuggestTooltipRenderOpts };
