+++ components/common/CharmEditor/components/inlinePalette/components/InlineCommandPalette.tsx
@@ -11,7 +11,7 @@ import type { ReactNode } from 'react';
 import { useMemo, useCallback, useEffect, useState } from 'react';
 
 import { useEditorViewContext } from '../../@bangle.dev/react/hooks';
-import { selectionTooltip } from '../../@bangle.dev/tooltip';
+import { hideSelectionTooltip } from '../../@bangle.dev/tooltip/selectionTooltip';
 import type { NestedPagePluginState } from '../../nestedPage/nestedPage.interfaces';
 import { GroupLabel } from '../../PopoverMenu';
 import { useInlinePaletteItems, useInlinePaletteQuery } from '../hooks';
@@ -135,7 +135,7 @@ export default function InlineCommandPalette({
 
   function closeSubMenu() {
     if (menuKey) {
-      selectionTooltip.hideSelectionTooltip(menuKey)(view.state, view.dispatch, view);
+      hideSelectionTooltip(menuKey)(view.state, view.dispatch, view);
     }
   }
