--- components/common/CharmEditor/components/inlinePalette/index.ts
@@ -1,1 +1,0 @@
-export * from './inlinePalette';
