+++ components/common/CharmEditor/components/linkedPage/linkedPage.specs.ts
@@ -2,7 +2,7 @@ import type { DOMOutputSpec } from '@bangle.dev/pm';
 import { log } from '@charmverse/core/log';
 
 import type { RawSpecs } from '../@bangle.dev/core/specRegistry';
-import * as suggestTooltip from '../@bangle.dev/tooltip/suggestTooltip';
+import { spec as SuggestTooltipSpec } from '../@bangle.dev/tooltip/suggestTooltipSpec';
 import { encloseNestedPage } from '../nestedPage/nestedPage.utils';
 
 import { linkedPageNodeName, linkedPageSuggestMarkName } from './linkedPage.constants';
@@ -47,7 +47,7 @@ export function linkedPageSpec(): RawSpecs {
         }
       }
     },
-    suggestTooltip.spec({ markName: linkedPageSuggestMarkName }),
+    SuggestTooltipSpec({ markName: linkedPageSuggestMarkName }),
     // add temporary stub to support old mark name
     {
       name: 'nestedPageSuggest',
