+++ lib/prosemirror/insertLinkedPage.ts
@@ -3,7 +3,7 @@ import { TextSelection } from '@bangle.dev/pm';
 import { rafCommandExec } from '@bangle.dev/utils';
 
 import { hideSuggestionsTooltip } from 'components/common/CharmEditor/components/@bangle.dev/tooltip/suggestTooltip';
-import { replaceSuggestionMarkWith } from 'components/common/CharmEditor/components/inlinePalette';
+import { replaceSuggestionMarkWith } from 'components/common/CharmEditor/components/inlinePalette/inlinePalette';
 import { linkedPageNodeName } from 'components/common/CharmEditor/components/linkedPage/linkedPage.constants';
 import { type LinkedPagePluginState } from 'components/common/CharmEditor/components/linkedPage/linkedPage.interfaces';
